A solution
It came out of nowhere, I was deeply comepllede to boot up the laptop and ship it. I will not merge this branch until it is fully tested,I was able to find a solution, not by using multi threading, which ended up not doing what I thought it would, but moveing away from haveing the checking happen in another function, I was stuck in this dilemma of how am I going to get a state change out of this function into the loop that is currently running. but for now I am happy.