This works
This is the distance formula without input filtering, so I put in a warning. I will merge this branch for now but no delete because I might revisit it to add input filtering. The reason I am not doing it now is because It would take a lot of code implement, but it will probably come about soon.